requirements.txt
numpy-financial==1.0.0 
plotly==5.9.0 
statsmodels==0.12.2
streamlit==1.11.1 
streamlit-folium==0.6.14 
streamlit-lottie==0.0.3
folium==0.12.1.post1
scikit-image==0.18.3 
geographiclib==1.52
geopy==2.2.0 
altair==4.2.0
tables==3.6.1 
